sm8150-common: thermal: Move BCL handling back to thermal engine
Due to nature of pixel thermal HAL (limited by the level of ThrottlingSeverity),
we can't really do this kind of throttling on the HAL since it can affect
overall system performance because the HAL also reports severity level to
the system, not just applying our cdev handling.

But sadly we can't really use BCL SoC directly for some reason, even
with user_space governor setup.

So in order to mitigate it, we have to introduce a dedicated zone for
monitoring the BCL percentage on the kernel device tree.

Co-authored-by: alanndz <alanndz7@gmail.com>
Signed-off-by: bagaskara815 <bagaskara815@gmail.com>